move googleLogin function to element | add googleLogin to Register screen
